팔로워 목록 조회 기능 추가

+++ src/main/java/com/brandpark/sharemusic/api/v1/account/AccountApiController.java
@@ -1,5 +1,7 @@
-package com.brandpark.sharemusic.api.v1.follow;
+package com.brandpark.sharemusic.api.v1.account;
 
+import com.brandpark.sharemusic.api.v1.account.dto.FollowerListPagingDto;
+import com.brandpark.sharemusic.api.v1.account.query.AccountQueryRepository;
 import com.brandpark.sharemusic.api.v1.exception.ApiException;
 import com.brandpark.sharemusic.api.v1.exception.Error;
 import com.brandpark.sharemusic.infra.config.auth.LoginAccount;
@@ -9,15 +11,18 @@ import com.brandpark.sharemusic.modules.account.domain.AccountRepository;
 import com.brandpark.sharemusic.modules.follow.Follow;
 import com.brandpark.sharemusic.modules.follow.FollowRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.web.PageableDefault;
 import org.springframework.web.bind.annotation.*;
 
 @RequestMapping("/api/v1")
 @RequiredArgsConstructor
 @RestController
-public class FollowApiController {
+public class AccountApiController {
 
     private final AccountRepository accountRepository;
     private final FollowRepository followRepository;
+    private final AccountQueryRepository accountQueryRepository;
 
     @PostMapping("/accounts/{targetId}/follow")
     public Long doFollow(@LoginAccount SessionAccount account, @PathVariable Long targetId) {
@@ -28,6 +33,10 @@ public class FollowApiController {
         Account target = accountRepository.findById(targetId)
                 .orElseThrow(() -> new ApiException(Error.NOT_FOUND_ACCOUNT_EXCEPTION));
 
+        if (followRepository.isFollowing(account.getId(), targetId)) {
+            throw new ApiException(Error.ILLEGAL_ACCESS_EXCEPTION);
+        }
+
         return followRepository.save(Follow.builder()
                 .follower(follower)
                 .target(target)
@@ -37,17 +46,25 @@ public class FollowApiController {
     @PostMapping("/accounts/{targetId}/unfollow")
     public Long doUnfollow(@LoginAccount SessionAccount account, @PathVariable Long targetId) {
 
-        Account unFollower = accountRepository.findById(account.getId())
-                .orElseThrow(() -> new ApiException(Error.NOT_FOUND_ACCOUNT_EXCEPTION));
+        if (!accountRepository.existsById(targetId)) {
+            throw new ApiException(Error.NOT_FOUND_ACCOUNT_EXCEPTION);
+        }
 
-        Account target = accountRepository.findById(targetId)
-                .orElseThrow(() -> new ApiException(Error.NOT_FOUND_ACCOUNT_EXCEPTION));
-
-        Follow follow = followRepository.findByFollowerIdAndTargetId(unFollower.getId(), target.getId())
+        Follow follow = followRepository.findByFollowerIdAndTargetId(account.getId(), targetId)
                 .orElseThrow(() -> new ApiException(Error.ILLEGAL_ACCESS_EXCEPTION));
 
         followRepository.delete(follow);
 
         return follow.getId();
+    }
+
+    @GetMapping("/accounts/{targetId}/followers")
+    public FollowerListPagingDto findAllFollowersByPaging(@PathVariable Long targetId, @PageableDefault(size = 6) Pageable pageable) {
+
+        if (accountRepository.existsById(targetId)) {
+            throw new ApiException(Error.NOT_FOUND_ACCOUNT_EXCEPTION);
+        }
+
+        return accountQueryRepository.findAllFollowerByPaging(targetId, pageable);
     }
 }
